base/requirements.txt
git+https://github.com/murmursh/browser-mod
python-dotenv
